Moved common code to abstract class `ArrayGrid`
