+++ src/core/brsTypes/nodes/MarkupList.ts
@@ -1,10 +1,18 @@
 import { FieldKind, FieldModel } from "./Field";
 import { AAMember } from "../components/RoAssociativeArray";
 import { ArrayGrid } from "./ArrayGrid";
-import { BrsInvalid, BrsString, BrsType, ContentNode, customNodeExists, Group, Int32, jsValueOf, rootObjects, ValueKind } from "..";
+import {
+    BrsInvalid,
+    BrsString,
+    BrsType,
+    customNodeExists,
+    Int32,
+    jsValueOf,
+    rootObjects,
+    ValueKind,
+} from "..";
 import { IfDraw2D, Rect } from "../interfaces/IfDraw2D";
 import { Interpreter } from "../../interpreter";
-import { rotateTranslation } from "../../scenegraph/SGUtil";
 import { BrsDevice } from "../../device/BrsDevice";
 
 export class MarkupList extends ArrayGrid {
@@ -18,7 +26,6 @@ export class MarkupList extends ArrayGrid {
     protected readonly footprintUri = "common:/images/focus_footprint.9.png";
     protected readonly margin: number;
     protected wrap: boolean;
-    protected hasNinePatch: boolean;
 
     constructor(initializedFields: AAMember[] = [], readonly name: string = "MarkupList") {
         super([], name);
@@ -88,24 +95,7 @@ export class MarkupList extends ArrayGrid {
         return this.handleUpDown(key);
     }
 
-    renderNode(interpreter: Interpreter, origin: number[], angle: number, draw2D?: IfDraw2D) {
-        if (!this.isVisible()) {
-            return;
-        }
-        const nodeTrans = this.getTranslation();
-        const drawTrans = angle !== 0 ? rotateTranslation(nodeTrans, angle) : nodeTrans.slice();
-        drawTrans[0] += origin[0];
-        drawTrans[1] += origin[1];
-        const size = this.getDimensions();
-        const rect = { x: drawTrans[0], y: drawTrans[1], width: size.width, height: size.height };
-        const rotation = angle + this.getRotation();
-        this.renderList(interpreter, rect, rotation, draw2D);
-        this.updateBoundingRects(rect, origin, rotation);
-        this.renderChildren(interpreter, drawTrans, rotation, draw2D);
-        this.updateParentRects(origin, angle);
-    }
-
-    protected renderList(
+    protected renderContent(
         interpreter: Interpreter,
         rect: Rect,
         rotation: number,
@@ -120,7 +110,7 @@ export class MarkupList extends ArrayGrid {
         const itemCompName = this.getFieldValue("itemComponentName") as BrsString;
         if (!customNodeExists(interpreter, itemCompName)) {
             BrsDevice.stderr.write(
-                `warning,[sg.markupgrid.create.fail] Failed to create markup item ${itemCompName}`
+                `warning,[sg.markuplist.create.fail] Failed to create markup item ${itemCompName}`
             );
             return;
         }
@@ -156,7 +146,7 @@ export class MarkupList extends ArrayGrid {
             if (index >= this.content.length) {
                 break;
             }
-            this.renderItem(interpreter, index, itemRect, rotation, draw2D);
+            this.renderItemComponent(interpreter, index, itemRect, rotation, draw2D);
             lastIndex = index;
             itemRect.x = rect.x;
             itemRect.y += itemRect.height + (rowSpacings[r] ?? spacing[1]);
@@ -165,55 +155,5 @@ export class MarkupList extends ArrayGrid {
         rect.y = rect.y - (this.hasNinePatch ? 4 : 0);
         rect.width = itemSize[0] + (this.hasNinePatch ? this.margin * 2 : 0);
         rect.height = displayRows * (itemSize[1] + (this.hasNinePatch ? 9 : 0));
-    }
-
-    protected renderItem(
-        interpreter: Interpreter,
-        index: number,
-        itemRect: Rect,
-        rotation: number,
-        draw2D?: IfDraw2D
-    ) {
-        const content = this.content[index];
-        if (!(content instanceof ContentNode)) {
-            return;
-        }
-        const nodeFocus = rootObjects.focused === this;
-        const focused = index === this.focusIndex;
-        if (!this.itemComps[index]) {
-            const itemComp = this.createItemComponent(interpreter, itemRect, content, focused);
-            if (itemComp instanceof Group) {
-                this.itemComps[index] = itemComp;
-            }
-        }
-        const drawFocus = jsValueOf(this.getFieldValue("drawFocusFeedback"));
-        const drawFocusOnTop = jsValueOf(this.getFieldValue("drawFocusFeedbackOnTop"));
-        if (focused && drawFocus && !drawFocusOnTop) {
-            this.renderFocus(itemRect, nodeFocus, draw2D);
-        }
-        const itemOrigin = [itemRect.x, itemRect.y];
-        this.itemComps[index].renderNode(interpreter, itemOrigin, rotation, draw2D);
-        if (focused && drawFocus && drawFocusOnTop) {
-            this.renderFocus(itemRect, nodeFocus, draw2D);
-        }
-    }
-
-    protected renderFocus(itemRect: Rect, nodeFocus: boolean, draw2D?: IfDraw2D) {
-        const focusBitmap = this.getBitmap("focusBitmapUri");
-        const focusFootprint = this.getBitmap("focusFootprintBitmapUri");
-        this.hasNinePatch = (focusBitmap?.ninePatch || focusFootprint?.ninePatch) === true;
-        const ninePatchRect = {
-            x: itemRect.x - 15,
-            y: itemRect.y - 15,
-            width: itemRect.width + 31,
-            height: itemRect.height + 30,
-        };
-        if (nodeFocus && focusBitmap) {
-            const rect = focusBitmap.ninePatch ? ninePatchRect : itemRect;
-            this.drawImage(focusBitmap, rect, 0, draw2D);
-        } else if (!nodeFocus && focusFootprint) {
-            const rect = focusFootprint.ninePatch ? ninePatchRect : itemRect;
-            this.drawImage(focusFootprint, rect, 0, draw2D);
-        }
     }
 }
